refactor(#1363): extract slack_runtime tests into module file
